+++ lockhart/cli/config.py
@@ -1,8 +1,8 @@
+from rich.console import Console
 import typer
 
 from lockhart.cli.common import verbose_callback
 from lockhart.config import config as configuration
-from lockhart.console import console
 
 config_app = typer.Typer()
 
@@ -26,4 +26,4 @@ def show(
         help="show the log messages",
     ),
 ):
-    console.print(configuration)
+    Console().print(configuration)