+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.6.2
+
+- fix: unable to edit tempfile on findows, 0.6.2.dev1
+- fix: `lockhart config show` does not print without --verbose, 0.6.2.dev2
+
+## 0.6.1
+
+- fix: toml items not string throw attribute error closes #34, 0.6.1.dev0
+- fix: tui does not work with empty history 0.6.1.dev0
+
 ## 0.6.0
 
 - fix: default to quiet console, make --verbose a common cli argument, #25 0.6.0.dev1